+++ modules/outbound-command-event-handler/src/domain/bulk_transaction_agg/handlers/index.ts
