+++ modules/private-shared-lib/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "rootDir": "./src",
     "outDir": "./dist",
-    "baseUrl": ".",
     "paths": {
       "@module-domain": [ "./src/domain/index.ts" ],
       "@module-errors": [ "./src/errors/index.ts" ],
