+++ package.json
@@ -21,10 +21,12 @@
     "modules/*"
   ],
   "scripts": {
+    "postinstall": "yarn husky install",
     "start": "nx run-many --output-style=stream --parallel --target=start",
     "start:api-svc": "yarn workspace @mojaloop/sdk-scheme-adapter-api-svc run start",
     "start:event-handler": "yarn workspace @mojaloop/sdk-scheme-adapter-event-handler run start",
     "start:command-handler": "yarn workspace @mojaloop/sdk-scheme-adapter-command-handler run start",
+    "build:no-cache": "nx run-many --all --target=build --skip-nx-cache",
     "build": "nx run-many --all --target=build",
     "build:affected": "nx affected --target=build",
     "build:openapi": "yarn workspace @mojaloop/sdk-scheme-adapter-api-svc run build:openapi",
